reshuffle ActionListFeatureConnected files to match existing "ConnectedComponent" paradigm.
