+++ packages/apollos-ui-connected/src/ActionListFeatureConnected/ActionListFeature/ActionListFeature.stories.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { storiesOf } from '@apollosproject/ui-storybook';
 
 import { BackgroundView, CenteredView } from '@apollosproject/ui-kit';
-import ActionListFeature from './ActionListFeature';
+import ActionListFeature from '.';
 
 const actions = [
   {
