+++ packages/apollos-ui-connected/src/ActionListFeatureConnected/ActionListFeature/ActionListFeature.tests.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import { Providers } from '../utils/testUtils';
+import { Providers } from '../../utils/testUtils';
 
-import ActionListFeature from './ActionListFeature';
+import ActionListFeature from '.';
 
 const actions = [
   {
