+++ packages/apollos-ui-connected/src/ActionListFeatureConnected/index.js
@@ -1,41 +1,6 @@
-import React from 'react';
-import { Query } from 'react-apollo';
-import { get } from 'lodash';
-import PropTypes from 'prop-types';
-import ActionListFeature from './ActionListFeature';
-import GET_ACTION_LIST_FEATURE from './getActionListFeature';
+import ActionListFeatureConnected from './ActionListFeatureConnected';
 
-function ActionListFeatureConnected({
-  featureId,
-  Component,
-  isLoading,
-  ...props
-}) {
-  return (
-    <Query query={GET_ACTION_LIST_FEATURE} variables={{ featureId }}>
-      {({ data, loading }) => (
-        <Component
-          {...get(data, 'node')}
-          {...props}
-          isLoading={loading || isLoading}
-        />
-      )}
-    </Query>
-  );
-}
+export ActionListFeature from './ActionListFeature';
+export GET_ACTION_LIST_FEATURE from './getActionListFeature';
 
-ActionListFeatureConnected.propTypes = {
-  Component: PropTypes.node,
-  featureId: PropTypes.string.isRequired,
-  isLoading: PropTypes.bool,
-};
-
-ActionListFeatureConnected.defaultProps = {
-  Component: ActionListFeature,
-};
-
-export {
-  ActionListFeatureConnected as default,
-  GET_ACTION_LIST_FEATURE,
-  ActionListFeature,
-};
+export default ActionListFeatureConnected;
